got the Listview to show images from the database

+++ app/src/main/java/com/example/mobilecoursework/model/CafeItem.kt
@@ -1,7 +1,7 @@
 package com.example.mobilecoursework.model
 
 class CafeItem(
-    var proName: String, var prodPrice: Double,
+    var proName: String, var prodPrice:Float,
     var prodImage: ByteArray,
     var prodAvailable: Boolean) {
 }
